requirements.txt
streamlit
pandas
matplotlib
streamlit-autorefresh
python-dotenv